requirements.txt
Django==4.2
psycopg2-binary==2.9.9
django-shortuuidfield==0.1.3
djangorestframework==3.14.0
pygments=2.16.1
